My own fetch Hook with loop solution
