+++ src/index.js
@@ -2,11 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "./styles/global-style.css";
 
-import Home from "./templates/Home";
+import { Home } from "./templates/Home";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(
-  <React.StrictMode>
-    <Home />
-  </React.StrictMode>,
-);
+root.render(<Home />);
