+++ src/templates/Home/index copy.jsx
@@ -1,0 +1,71 @@
+import React, { useEffect, useState, useCallback } from "react";
+import { loadPosts } from "../../utils/load-posts";
+import { Posts } from "../../components/Posts";
+import { Button } from "../../components/Button";
+import "./styles.css";
+import { TextInput } from "../../components/TextInput";
+
+export const Home = () => {
+  const [posts, setPosts] = useState([]);
+  const [allPosts, setAllPosts] = useState([]);
+  const [page, setPage] = useState(0);
+  const [postsPerPage] = useState(2);
+  const [searchValue, setSearchValue] = useState("");
+
+  const noMorePosts = page + postsPerPage >= allPosts.length;
+
+  const filteredPosts = searchValue
+    ? allPosts.filter((post) => {
+        return post.title.toLowerCase().includes(searchValue.toLowerCase());
+      })
+    : posts;
+
+  const handleLoadPosts = useCallback(async (page, postsPerPage) => {
+    const postsAndPhotos = await loadPosts();
+    setPosts(postsAndPhotos.slice(page, postsPerPage));
+    setAllPosts(postsAndPhotos);
+  }, []);
+
+  useEffect(() => {
+    console.log(new Date().toLocaleString("pt-br"));
+    handleLoadPosts(0, postsPerPage);
+  }, [handleLoadPosts, postsPerPage]);
+
+  const loadMorePosts = () => {
+    const nextPage = page + postsPerPage;
+
+    setPosts([...posts, ...allPosts.slice(nextPage, nextPage + postsPerPage)]);
+    setPage(nextPage);
+  };
+
+  const handleChange = (e) => {
+    const { value } = e.target;
+    setSearchValue(value);
+  };
+
+  return (
+    <section className="container">
+      <div className="search-container">
+        {!!searchValue && <h1>Search Value: {searchValue}</h1>}
+
+        <TextInput searchValue={searchValue} handleChange={handleChange} />
+      </div>
+
+      {filteredPosts.length === 0 && <p>Não existem posts</p>}
+
+      <Posts posts={filteredPosts} />
+
+      <div className="button-container">
+        {!searchValue && (
+          <Button
+            onClick={loadMorePosts}
+            text="Load more posts"
+            disabled={noMorePosts}
+          />
+        )}
+      </div>
+    </section>
+  );
+};
+
+export default Home;
